Create Program-15.java
Create two subclasses of Shape called “Rectangle” and 
“Triangle” with the following properties and methods: 
Rectangle: 
 A private double called “length” 
 A private double called “width” 
 A constructor that takes in two doubles (length 
and width) and sets the corresponding fields. 
 An implementation of the calculateArea() 
method that return length * width. 
Triangle: 
 A private double called “base”
 A private double called “height”
 A constructor that takes in two doubles (base 
and height) and sets the corresponding fields.
 An implementation of the calculateArea() 
method that returns (base * height)/2.
Create a main class called “ShapeTester” with the 
following properties and methods: 
 A static method called “printShapeArea” that 
takes in a Shape object and prints out the area of 
the shape. 
 In the main method, create one Rectangle and 
one Triangle object, pass them into the 
printShapeArea method, and observe the output.