Create PackageTester.java
A class called “Person” with the following properties and 
methods: 
 A private String called “name” 
 A private int called “age” 
 A constructor that takes in a name and age and 
sets the corresponding fields. 
 Getter and setters for the name an
A class called “Student” that extends the Person class 
and has the following properties and methods: 
 A private int called “id” 
 A private String called “major” 
 A constructor that takes in a name, age, id, and 
major and sets the corresponding fields using the 
super () constructor. 
 Getter and setters for the id and major fields. 
 A toString() method that calls the 
super.toString() method and adds the subject 
field. 
Create a main class called “PackageTester” in a different 
package that does the following: 
 Creates one Student object and one Teacher 
object using the constructors provided. 
 Uses the setters to chage the id, major, and 
subject fields of the objects. 
 Print out the string representations of the objects using the toString() method.